linux arm64, nolapack compile fixes

+++ 2.0/libdeflate/lib/arm/arm_cpu_features.c
@@ -81,7 +81,7 @@ static void scan_auxv(unsigned long *hwcap, unsigned long *hwcap2)
 				goto out;
 			}
 			filled += ret;
-		} while (filled < 2 * sizeof(long));
+		} while ((size_t)filled < 2 * sizeof(long));
 
 		i = 0;
 		do {
@@ -94,7 +94,7 @@ static void scan_auxv(unsigned long *hwcap, unsigned long *hwcap2)
 				*hwcap2 = value;
 			i += 2;
 			filled -= 2 * sizeof(long);
-		} while (filled >= 2 * sizeof(long));
+		} while ((size_t)filled >= 2 * sizeof(long));
 
 		memmove(auxbuf, &auxbuf[i], filled);
 	}
